Attempting Context Locals!
We need a way to pass settings back from Ambidex into the routes.  Since
modules in Node can't close over one another, I tried to fake it by
abusing the error system to implement context locals.  Unfortunately,
the stack expires before any asynchronous code is run.

I'm committing this for posterity, and to bookmark it in case I need
this solution later.

As for passing the settings down, I need a new strategy.  The most
likely candidate is to create AmbidexRoutes, which wouuld inject a
settings prop onto every route it's passed, then wrap them in <Routes>
and return that.  Not sure yet if it'll work - I should talk to
@mjackson about it.